fix(vector.nix): Make GCP metadata fetching more robust by adding fallbacks to 'unknown'
This change ensures that if the curl commands to fetch GCP metadata fail, the environment variables will be set to 'unknown' instead of causing the service to fail. This enhances the stability of the vector agent startup.

Additionally, prefixed the `EnvironmentFile` path with a hyphen (`-`) to treat the file as optional. This prevents the service from failing to start if the `/run/vector-gcp-env` file is not present for any reason.