feat(gcp/pulumi-infra): Add logging service account and key creation
This commit introduces the creation of a dedicated service account for logging purposes in Google Cloud Platform.
This service account is granted the `roles/logging.logWriter` role, allowing it to send logs to Cloud Logging.
A private key in the Google Credentials File format is also generated for this service account, which will be used by the VM instance to authenticate with Cloud Logging.
The email of the service account and its private key are exported for use in other parts of the infrastructure or application.