refactor(config): update GCP project ID placeholder in Pulumi.example.yaml and mise.toml
The GCP project ID placeholder in `Pulumi.example.yaml` has been changed from `dev-vm-provisioning` to `GCP_PROJECT_ID_PLACEHOLDER`.
This change is to make it clear that this value should be replaced with the actual GCP project ID.
The `mise.toml` file has also been updated to reflect this change in the `run` command's `perl` substitution.