refactor(mise.toml): improve pulumi stack initialization by using sed to replace placeholder values and mv to rename files
This change aims to streamline the process of initializing a new Pulumi stack. Previously, the workflow involved copying an example Pulumi configuration file and then using `perl` to replace a placeholder. This commit introduces `sed` for direct in-place replacement of the placeholder in `Pulumi.example.yaml`.

Additionally, the file renaming logic has been refined. Instead of copying and then modifying, the `mv` command is now used to rename `Pulumi.example.yaml` to `Pulumi.<hostname>.yaml`. This reduces the number of file operations and ensures that the original `Pulumi.example.yaml` is restored after the renaming, maintaining a cleaner state.