refactor(fish/config.fish): Update gitconfig alias path and add verbosity to chezmoi apply alias
The path for the gitconfig alias `eg` was incorrect, pointing to a non-existent directory. This commit corrects the path to `$HOME/code/nixos-dotfiles/chezmoi/dot_gitconfig`.

Additionally, the `chezmoi apply` alias `cma` is updated to include the `-v` flag for verbose output. This change aims to provide more detailed feedback during the apply process, improving the user experience when managing dotfiles with chezmoi.