refactor(git): update git config alias to point to the correct tmpl file
The alias `eg` was previously pointing to `dot_gitconfig`. This commit updates it to correctly point to `dot_gitconfig.tmpl` which is the actual template file used by Chezmoi for generating the git configuration. This ensures that changes to the git configuration via the alias are made to the correct source file.

Also, the shebang line in `dot_gitconfig.tmpl` is updated to include `ft=gitconfig` for better syntax highlighting in Vim when editing the file.