docs(GCP_USAGE.md): add github personal access token to secrets instructions to facilitate automated deployment
This change updates the `GCP_USAGE.md` file to include instructions for generating a GitHub Personal Access Token (PAT) and adding it to the secrets file. This is necessary for recent changes in the project that require a GitHub PAT for automated deployments to Google Cloud Engine instances. The `gcp_coral.yaml` and `gcp_example.yaml` files have also been updated to include the new `githubClassicPersonalAccessToken` field, encrypted with SOPS.