feat(modules/nixos-dotfiles): add service to clone nixos-dotfiles repository
This commit introduces a new systemd service that automatically clones
the nixos-dotfiles repository to the user's home directory upon system
startup. This ensures that the dotfiles are always available and in sync
with the repository.

The service depends on the `add-ssh-key-to-github.service`, guaranteeing
that SSH access to GitHub is established before attempting to clone.
It also ensures that the `code` directory exists and creates it if
necessary. If the repository already exists, the service exits successfully
without performing the clone again.