refactor(interactive-cli.nix): move polkit configuration to user module
The previous configuration for polkit rules was located in the `interactive-cli.nix` module. These rules are user-specific, allowing members of the wheel group to reboot or shutdown without a password. To better organize the configuration, this functionality has been moved to the `user.nix` module, enhancing the clarity and maintainability of the NixOS configuration.