refactor(interactive-cli.nix): move chezmoi and mise to unstable packages section
This change moves the `chezmoi` and `mise` packages to the `pkgs-pinned-unstable` section. This is done to potentially leverage newer versions of these tools more readily, as they are frequently updated and offer new features.Keeping them with the other unstable packages simplifies management and ensures consistency in how frequently updated tools are handled within the system configuration.