feat(configuration): enable SQL tools option
Adds a new boolean option `my.enableSqlTools` to the NixOS configuration.
This option allows users to easily enable or disable the installation of SQL-related tools.

This change also integrates the new option into the package
management system by adding `sql-studio` to the list of packages
installed when `my.enableSqlTools` is true.
This provides a convenient way to manage SQL client installations
through the NixOS configuration.