feat(chezmoi): add KeepassXC database configuration and Intelephense license template
Adds the KeepassXC database configuration to the `.chezmoi.toml.tmpl` file, pointing to the encrypted database in the `destDir`. This allows chezmoi to manage secrets encrypted with KeepassXC.

Also adds a new template file `intelephense/licence.txt.tmpl` that retrieves the Intelephense license key from KeepassXC using the `keepassxc` function. This automates the process of keeping the Intelephense license information up-to-date.