feat(chezmoi): add chezmoi-vim plugin configuration
This commit adds the configuration for the chezmoi-vim plugin, which allows Neovim to effectively manage dotfiles.

It also includes a minor change to the fish config to remove an unnecessary comment.

Finally, it updates the treesitter configuration to enable syntax highlighting for chezmoitmpl files,
and removes old vim: ft= syntax highlighting from templates.