refactor(chezmoi-vim.lua): use .chezmoi.sourceDir template for chezmoi#source_dir_path
This change updates the `chezmoi#source_dir_path` setting in the Neovim configuration to use the `{{ .chezmoi.sourceDir }}` template variable. This will ensure that chezmoi automatically uses the correct source directory, regardless of where the configuration files are located or how chezmoi is invoked. Previously, the path was hardcoded, which could lead to issues if the directory structure changed.