feat(configuration): add option to enable windows tools and dos2unix
This commit introduces a new option, `my.enableWindowsTools`, to control the
installation of Windows-related utilities. It also includes `dos2unix` in the
list of optional packages when this new option is enabled. This change aims
to provide better compatibility and convenience for users who might need to
work with files originating from Windows environments.