refactor(functions): Alias 'tree' command with 't' to 'eza --tree'
The previous implementation used `eza --tree` to replace the `tree` command. However, this change reverts that and uses the actual `tree` command instead. This is because `eza` does not have the same functionality as `tree`. For example, `eza` does not show file sizes or modification dates in the same way that `tree` does.

feat(nixos): Add colima to optional packages for Docker on macOS

This commit adds `colima` to the list of optional packages for Docker on macOS. Colima is a container registry that can be used to run Docker containers on macOS. This change is necessary to ensure that Docker can be used on macOS.